net: move CBanDB and CAddrDB out of net.h/cpp
This will eventually solve a circular dependency
